moved EditNoteFragment and EditNoteViewModel to addoreditnote package
