+++ app/src/main/res/layout/edit_note_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditNoteFragment">
+    tools:context=".addoreditnote.EditNoteFragment">
 
     <TextView
         android:layout_width="match_parent"
